+++ src/main/java/com/travelservice/domain/admin/dto/user/UserResponseDto.java
@@ -1,4 +1,6 @@
-package com.travelservice.domain.admin.dto;
+package com.travelservice.domain.admin.dto.user;
+
+import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,9 +11,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class AdminUserResponseDto {
+public class UserResponseDto {
 	private Long userId;
 	private String name;
 	private String email;
+	private String phoneNumber;
 	private Integer roleCode;
+	private LocalDateTime createAt;
+	private LocalDateTime deleteAt;
 }
