+++ src/main/java/com/travelservice/domain/admin/dto/user/UserUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.travelservice.domain.admin.dto;
+package com.travelservice.domain.admin.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
