+++ src/main/java/com/travelservice/domain/admin/service/AdminUserService.java
@@ -9,11 +9,11 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.travelservice.domain.admin.dto.AdminUserResponseDto;
-import com.travelservice.domain.admin.dto.PagedAdminUserResponseDto;
-import com.travelservice.domain.admin.dto.PagedUserResponseDto;
-import com.travelservice.domain.admin.dto.UserResponseDto;
-import com.travelservice.domain.admin.dto.UserUpdateRequestDto;
+import com.travelservice.domain.admin.dto.user.AdminUserResponseDto;
+import com.travelservice.domain.admin.dto.user.PagedAdminUserResponseDto;
+import com.travelservice.domain.admin.dto.user.PagedUserResponseDto;
+import com.travelservice.domain.admin.dto.user.UserResponseDto;
+import com.travelservice.domain.admin.dto.user.UserUpdateRequestDto;
 import com.travelservice.domain.admin.repository.AdminUserRepository;
 import com.travelservice.domain.user.entity.User;
